requirements.txt
beautifulsoup4
fastapi
httpx
langchain-ollama
langgraph
python-dotenv
python-telegram-bot
tavily-python
uvicorn
